+++ Sources/Userlog/userlog/lib/features/about/screens/about_screen.dart
@@ -1,9 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_svg/flutter_svg.dart';
 
-import 'package:userlog/common/styles/fonts.dart';
-import 'package:userlog/features/about/assets/about_images.dart';
-import 'package:userlog/features/about/assets/about_texts.dart';
+import 'package:userlog/features/about/widgets/about_widgets/about_app_widget.dart';
+import 'package:userlog/features/about/widgets/about_widgets/about_center_widget.dart';
 
 class AboutScreen extends StatelessWidget {
   const AboutScreen({super.key});
@@ -40,58 +38,14 @@ class AboutScreen extends StatelessWidget {
             height: sizeBoxHeightInputPadding,
             width: screenWidth,
           ),
-          Container(
-            margin: const EdgeInsets.only(left: 0),
-            padding: EdgeInsets.only(right: sizeBoxWidthInputPadding),
-            child: Row(
-              mainAxisSize: MainAxisSize.min,
-              children: <Widget>[
-                IconButton(
-                  icon: Align(
-                    alignment: Alignment.bottomLeft,
-                    child: SvgPicture.asset(
-                      AboutImages.exitButtonImage,
-                      fit: BoxFit.cover,
-                    ),
-                  ),
-                  iconSize: 51,
-                  onPressed: () {
-                    Navigator.pop(context);
-                  },
-                ),
-                const Text(
-                  AboutTexts.aboutText,
-                  style: TextStyle(
-                    color: Colors.white,
-                    fontSize: 42,
-                    fontWeight: FontWeight.bold,
-                    fontFamily: Fonts.mainFontFamily,
-                  ),
-                ),
-              ],
-            ),
-          ),
+          AboutAppWidget(sizeBoxWidthInputPadding: sizeBoxWidthInputPadding),
           SizedBox(
             height: sizeBoxHeightInputPadding,
             width: screenWidth,
           ),
-          Center(
-            child: Container(
-              padding: EdgeInsets.fromLTRB(
-                  aboutUsTextWidthInputPadding,
-                  aboutUsTextHeightInputPadding,
-                  aboutUsTextWidthInputPadding,
-                  0),
-              child: const Text(
-                AboutTexts.aboutUsText,
-                style: TextStyle(
-                  color: Colors.white60,
-                  fontSize: 20,
-                  fontFamily: Fonts.mainFontFamily,
-                ),
-              ),
-            ),
-          ),
+          AboutCenterWidget(
+              aboutUsTextWidthInputPadding: aboutUsTextWidthInputPadding,
+              aboutUsTextHeightInputPadding: aboutUsTextHeightInputPadding),
         ],
       ),
     );
